move MerkleTree.spec.ts
