+++ scripts/helpers/utils.ts
@@ -46,37 +46,6 @@ export function getZeroHash(zeroValue: any) {
     return ethers.utils.keccak256(abiCoder.encode(["uint256"], [zeroValue]));
 }
 
-export async function getMerkleRootFromLeaves(
-    dataLeaves: string[],
-    maxDepth: number
-) {
-    let nodes: string[] = dataLeaves.slice();
-    const defaultHashesForLeaves: string[] = defaultHashes(maxDepth);
-    let odd = nodes.length & 1;
-    let n = (nodes.length + 1) >> 1;
-    let level = 0;
-    while (true) {
-        let i = 0;
-        for (; i < n - odd; i++) {
-            let j = i << 1;
-            nodes[i] = getParentLeaf(nodes[j], nodes[j + 1]);
-        }
-        if (odd == 1) {
-            nodes[i] = getParentLeaf(
-                nodes[i << 1],
-                defaultHashesForLeaves[level]
-            );
-        }
-        if (n == 1) {
-            break;
-        }
-        odd = n & 1;
-        n = (n + 1) >> 1;
-        level += 1;
-    }
-    return nodes[0];
-}
-
 export function sign(signBytes: string, wallet: Wallet) {
     const h = ethUtils.toBuffer(signBytes);
     const signature = ethUtils.ecsign(h, wallet.getPrivateKey());
